feat(#718): Implements Healthcheck port resolution for Thorntail
