+++ core/src/test/resources/util/project-default.yml
@@ -1,0 +1,3 @@
+thorntail:
+  http:
+    port: 8082
