+++ enricher/fabric8/src/test/resources/project-defaults.yml
@@ -1,0 +1,3 @@
+thorntail:
+  http:
+    port: 8082
